Added RO-Access to Name Attribute
If quick paging through filenames using ListObjects and ListCurrentObjects, the current model requires the attributes to be requested from the Backblaze API.

However, the object already contains it's name, so this roundtrip is unnecessary. The function Object.Name() allows to quickly access the name without having to call back to the Backblaze API once per object, saving a couple hundred API calls in the process.