pyre: copy and paste a bunch of json code
In order to remain compatible with the B2 API, our emitted json needs to
send ints as ints.  Unfortunately, the proto3 spec says that 64 bit ints
(and uints) should be sent as strings.

In order to "correctly" marshal int64 and uint64 as ints, this adds a
custom marshaler to the API service.