Expose contentMd5 field in GetFileInfoResponse.
It's 2020 and we're adding an MD5 hash, you gotta be kidding me, right?
Technically yes, however the use case behind this is that AWS S3 uses
an MD5 hash as ETag for historical reasons. In order to ease working
with client libraries that rely on this (not officially deprecated)
behaviour when accessing B2 through an S3 compatibility layer such as
MinIO Gateway, it would be nice to expose the field.

Quoting the B2 docs:
> contentMd5 [optional]
> The MD5 of the bytes stored in the file as a 32-digit hex string.
> Not all files have an MD5 checksum, so this field is optional,
> and set to null for files that do not have one. Large files do not
> have MD5 checksums, and the value is null.